# feat: Use indexes for findWithNewOrChanged