--- a/packages/orm/src/EntityManager.ts
+++ b/packages/orm/src/EntityManager.ts
@@ -201,12 +201,13 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
   /** When we're flushing, the connection/transaction. */
   public txn: TX | undefined;
   public entityLimit: number = defaultEntityLimit;
-  readonly #entities: Entity[] = [];
+  readonly #entitiesByTag: Map<string, Entity[]> = new Map();
+  readonly #entitiesArray: Entity[] = [];
   // Indexes the currently loaded entities by their tagged ids. This fixes a real-world
   // performance issue where `findExistingInstance` scanning `#entities` was an `O(n^2)`.
   readonly #entityIndex: Map<string, Entity> = new Map();
   // Provides field-based indexing for entity types with >1000 entities to optimize findWithNewOrChanged
-  readonly #indexManager: IndexManager = new IndexManager();
+  readonly #indexManager = new IndexManager();
   #isValidating: boolean = false;
   readonly #pendingChildren: Map<string, Map<string, { adds: Entity[]; removes: Entity[] }>> = new Map();
   #preloadedRelations: Map<string, Map<string, Entity[]>> = new Map();
@@ -271,10 +272,8 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
       mutatedCollections: new Set(),
       hooks: this.#hooks,
       rm: this.#rm,
+      indexManager: this.#indexManager,
       isLoadedCache: this.#isLoadedCache,
-      get indexManager() {
-        return em.#indexManager;
-      },
 
       joinRows(m2m: ManyToManyCollection<any, any>): JoinRows {
         return getOrSet(em.#joinRows, m2m.joinTableName, () => new JoinRows(m2m, em.#rm));
@@ -312,7 +311,7 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
 
   /** Returns a read-only shallow copy of the currently-loaded entities. */
   get entities(): ReadonlyArray<Entity> {
-    return [...this.#entities];
+    return [...this.#entitiesArray];
   }
 
   /** Looks up `id` in the list of already-loaded entities. */
@@ -554,7 +553,9 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
     // clause before knowing if it should adjust teh amount.
     const isSelectAll = Object.keys(where).length === 0;
     if (isSelectAll) {
-      for (const entity of this.#entities) {
+      const tagged = this.#entitiesByTag.get(getMetadata(type).tagName) ?? [];
+      for (const entity of tagged) {
+        // Still do an `instanceof` to handle subtypes
         if (entity instanceof type) {
           if (entity.isNewEntity) {
             count++;
@@ -583,16 +584,16 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
    * @param type the entity type to find
    * @param where the fields to look up the existing entity by
    */
-  async findWithNewOrChanged<T extends EntityW, F extends Partial<OptsOf<T>>>(
+  async findWithNewOrChanged<T extends Entity, F extends Partial<OptsOf<T>>>(
     type: EntityConstructor<T>,
     where: F,
   ): Promise<T[]>;
-  async findWithNewOrChanged<T extends EntityW, F extends Partial<OptsOf<T>>, const H extends LoadHint<T>>(
+  async findWithNewOrChanged<T extends Entity, F extends Partial<OptsOf<T>>, const H extends LoadHint<T>>(
     type: EntityConstructor<T>,
     where: F,
     options?: { populate?: H; softDeletes?: "include" | "exclude" },
   ): Promise<Loaded<T, H>[]>;
-  async findWithNewOrChanged<T extends EntityW, F extends Partial<OptsOf<T>>, const H extends LoadHint<T>>(
+  async findWithNewOrChanged<T extends Entity, F extends Partial<OptsOf<T>>, const H extends LoadHint<T>>(
     type: EntityConstructor<T>,
     where: F,
     options?: { populate?: H; softDeletes?: "include" | "exclude" },
@@ -603,7 +604,7 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
     const copy: any = Object.fromEntries(Object.entries(where).filter(([, v]) => v !== undefined));
     const persisted = await this.find(type, copy, { softDeletes });
     const unchanged = persisted.filter((e) => !e.isNewEntity && !e.isDirtyEntity && !e.isDeletedEntity);
-    const maybeNew = this.#findEntitiesWithIndexes(type, where).filter(
+    const maybeNew = this.#filterEntities<T>(type, where).filter(
       (e) => (e.isNewEntity || e.isDirtyEntity) && !e.isDeletedEntity,
     );
     const found = [...unchanged, ...maybeNew];
@@ -1207,6 +1208,9 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
 
   /** Registers a newly-instantiated entity with our EntityManager; only called by entity constructors. */
   register(entity: Entity, maybeExplicitId?: string): void {
+    const baseMeta = getBaseMeta(getMetadata(entity));
+
+    // Keep our indexes up to date...
     const maybeId = entity.idTaggedMaybe ?? maybeExplicitId;
     if (maybeId) {
       if (this.findExistingInstance(maybeId) !== undefined) {
@@ -1214,18 +1218,20 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
       }
       this.#entityIndex.set(maybeId, entity);
     }
+    this.#entitiesArray.push(entity);
+    const set = this.#entitiesByTag.get(baseMeta.tagName) ?? [];
+    if (set.length === 0) this.#entitiesByTag.set(baseMeta.tagName, set);
+    set.push(entity);
 
-    this.#entities.push(entity);
-    if (this.#entities.length >= this.entityLimit) {
+    if (this.#entitiesArray.length >= this.entityLimit) {
       throw new Error(`More than ${this.entityLimit} entities have been instantiated`);
     }
 
-    // Check if we should enable indexing for this entity type
-    this.#maybeEnableIndexingForEntityType(entity);
+    // If indexing is enabled for this type, add it...
+    this.#indexManager.maybeIndexEntity(entity);
 
     // Set a default createdAt/updatedAt that we'll keep if this is a new entity, or over-write if we're loaded an existing row
     if (entity.isNewEntity) {
-      const baseMeta = getBaseMeta(getMetadata(entity));
       const { createdAt, updatedAt } = baseMeta.timestampFields ?? {};
       const { data } = getInstanceData(entity);
       const now = new Date();
@@ -1560,7 +1566,13 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
     this.#preloadedRelations = new Map();
     const deepLoad = param && "deepLoad" in param && param.deepLoad;
     let todo =
-      param === undefined ? this.#entities : Array.isArray(param) ? param : isEntity(param) ? [param] : this.#entities;
+      param === undefined
+        ? this.#entitiesArray
+        : Array.isArray(param)
+          ? param
+          : isEntity(param)
+            ? [param]
+            : this.#entitiesArray;
     const done = new Set<Entity>();
     while (todo.length > 0) {
       const copy = [...todo];
@@ -1672,10 +1684,13 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
 
         // This is a mini copy-paste of em.register that doesn't re-findExistingInstance
         this.#entityIndex.set(taggedId, entity as any);
-        this.#entities.push(entity as any);
-        if (this.#entities.length >= this.entityLimit) {
+        this.#entitiesArray.push(entity as any);
+        if (this.#entitiesArray.length >= this.entityLimit) {
           throw new Error(`More than ${this.entityLimit} entities have been instantiated`);
         }
+        const set = this.#entitiesByTag.get(maybeBaseMeta.tagName) ?? [];
+        if (set.length === 0) this.#entitiesByTag.set(maybeBaseMeta.tagName, set);
+        set.push(entity as any);
       } else if (options?.overwriteExisting === true) {
         // Usually if the entity already exists, we don't write over it, but in this case we assume that
         // `EntityManager.refresh` is telling us to explicitly load the latest data.
@@ -1863,42 +1878,16 @@ export class EntityManager<C = unknown, Entity extends EntityW = EntityW, TX ext
     await this.flush();
   }
 
-  /**
-   * Checks if indexing should be enabled for the entity type and enables it if threshold is reached.
-   */
-  #maybeEnableIndexingForEntityType(entity: Entity): void {
-    const meta = getMetadata(entity);
-    const entityType = meta.cstr as any;
-    const typeName = meta.type;
-
-    // Count entities of this type
-    const entitiesOfType = this.#entities.filter((e) => e instanceof entityType);
-
-    // If not indexed yet, check if we should enable indexing
-    if (!this.#indexManager.indexedTypes.has(typeName)) {
-      if (this.#indexManager.shouldIndexType(entitiesOfType.length)) {
-        // enableIndexingForType will index all entities including the current one
-        this.#indexManager.enableIndexingForType(entityType, entitiesOfType);
-        return; // No need to add individually since enableIndexingForType handles it
-      }
+  /** Returns entities matching the filter using indexed search when available. */
+  #filterEntities<T extends Entity>(cstr: EntityConstructor<T>, where: any): T[] {
+    const meta = getMetadata(cstr);
+    const entities = (this.#entitiesByTag.get(meta.tagName) as T[]) ?? [];
+    if (this.#indexManager.shouldIndexType(entities.length)) {
+      this.#indexManager.enableIndexingForType(meta, entities);
+      return this.#indexManager.findMatching(meta, where)!.filter((e) => !e.isDeletedEntity);
     } else {
-      // Type is already indexed, add this entity to the index
-      this.#indexManager.addEntity(entity);
+      return this.entities.filter((e) => e instanceof cstr && !e.isDeletedEntity && entityMatches(e, where)) as T[];
     }
-  }
-
-  /**
-   * Returns entities matching the filter using indexed search when available.
-   */
-  #findEntitiesWithIndexes<T extends Entity>(entityType: any, where: any): T[] {
-    // Try indexed search first
-    const indexedResults = this.#indexManager.findMatching(entityType, where);
-    if (indexedResults !== null) {
-      return indexedResults as T[];
-    }
-
-    // Fallback to linear search
-    return this.entities.filter((e) => e instanceof entityType && !e.isDeletedEntity && entityMatches(e, where)) as T[];
   }
 }
 
@@ -1918,12 +1907,12 @@ export interface EntityManagerInternalApi {
 
   hooks: Record<EntityManagerHook, HookFn<any>[]>;
   rm: ReactionsManager;
+  indexManager: IndexManager;
   preloader: PreloadPlugin | undefined;
   isValidating: boolean;
   checkWritesAllowed: () => void;
   get fieldLogger(): FieldLogger | undefined;
   get isLoadedCache(): IsLoadedCache;
-  get indexManager(): IndexManager;
 }
 
 export function getEmInternalApi(em: EntityManager): EntityManagerInternalApi {