--- a/packages/tests/integration/src/EntityManager.factories.test.ts
+++ b/packages/tests/integration/src/EntityManager.factories.test.ts
@@ -29,8 +29,8 @@ import {
   SmallPublisher,
 } from "@src/entities";
 import { isPreloadingEnabled, newEntityManager, queries, resetQueryCount } from "@src/testEm";
-import { maybeNew, maybeNewPoly, newTestInstance, noValue, setFactoryWriter, testIndex } from "joist-orm";
 import ansiRegex from "ansi-regex";
+import { maybeNew, maybeNewPoly, newTestInstance, noValue, setFactoryWriter, testIndex } from "joist-orm";
 
 let factoryOutput: string[] = [];