I didn't really expect to "need in-memory indexing" for Joist, but we've had several lifecycle hooks (like `beforeFlush`) get invoked for ~1,000+ entities (in a large, bulk upload/bulk update), which if the lifecycle does an innocent-looking `em.findWithNewOrChanged`, would turn into a `O(N^2)`.

I.e. 1,000 entities all doing a linear-scan for "did any of the other 1,000 entities `entityMatches(...my find argument...)`".

We can avoid the `O(N^2)` by reverse indexing the field values of the 1,000 entities, and then doing map/set lookups for `entityMatches` instead of a linear scan. 

This test:

```ts
      const em = newEntityManager();
      const n = 5_000;
      // Create entities to trigger indexing
      zeroTo(n).forEach((i) => em.create(Author, { firstName: `Author${i}` }));
      // Test that searches are now efficient
      const start = performance.now();
      const results = await Promise.all(
        zeroTo(n).map((i) => em.findWithNewOrChanged(Author, { firstName: `Author${i}` })),
      );
      const end = performance.now();
```

Without indexing takes 7.5 seconds.

With indexing it takes 314 ms, which is a 23x speedup.

Even with `n=500`, unindexed takes 100ms and indexes takes 50ms. So I'm going to set the threshold at 500.

To avoid the overhead of indexing entities unnecessarily, we only enable the index when:

- There is at least 500 instances of a given entity type in memory, and
- `findWithNewOrChanged` (or `findOrCreate` in a follow-up PR) is called

...technically if we have 1,000 instances, and there is only _one_ call to `findWithNewOrChanged`, we have probably indexed unnecessarily, so we could potentially count the number of `findWithNewOrChanged` invocations and only index after ~2-3 calls, i.e. as a heuristic to guess "we're probably in an `O(n^2)`".